fix load config
