+++ main.go
@@ -2,11 +2,7 @@ package main
 
 import (
 	"embed"
-	"encoding/json"
-	"fmt"
-	"os"
 
-	"github.com/adrg/xdg"
 	"github.com/wailsapp/wails/v2"
 	"github.com/wailsapp/wails/v2/pkg/options"
 	"github.com/wailsapp/wails/v2/pkg/options/assetserver"
@@ -16,22 +12,12 @@ import (
 var assets embed.FS
 
 func main() {
-	configPath, err := xdg.ConfigFile("njournal/config.json")
-	if err != nil {
-		fmt.Println("Error getting config path:", err)
-		return
-	}
-
-	appConfig, err := loadConfig(configPath)
-
 	// Create an instance of the app structure
-	app := NewApp(appConfig)
+	app := NewApp()
 
 	// Create application with options
-	err = wails.Run(&options.App{
-		Title:  "njournal",
-		Width:  appConfig.Width,
-		Height: appConfig.Height,
+	err := wails.Run(&options.App{
+		Title: "njournal",
 		AssetServer: &assetserver.Options{
 			Assets: assets,
 		},
@@ -47,56 +33,4 @@ func main() {
 	if err != nil {
 		println("Error:", err.Error())
 	}
-}
-
-type AppConfig struct {
-	Width        int    `json:"width"`
-	Height       int    `json:"height"`
-	DataBasePath string `json:"databasePath"`
-}
-
-func saveConfig(configPath string, appConfig AppConfig) error {
-	bytes, err := json.Marshal(appConfig)
-	if err != nil {
-		return fmt.Errorf("Error marshalling config: %s", err)
-	}
-
-	err = os.WriteFile(configPath, bytes, 0644)
-	if err != nil {
-		return fmt.Errorf("Error writing config: %s", err)
-	}
-
-	return nil
-}
-
-func loadConfig(configPath string) (*AppConfig, error) {
-	bytes, err := os.ReadFile(configPath)
-	if _, ok := err.(*os.PathError); ok {
-		// Create the config file
-		dataPath, err := xdg.DataFile("njournal/data/njournal.db")
-
-		if err != nil {
-			return nil, fmt.Errorf("Error getting data path: %s", err)
-		}
-
-		appConfig := AppConfig{
-			Width:        800,
-			Height:       600,
-			DataBasePath: dataPath,
-		}
-
-		if err := saveConfig(configPath, appConfig); err != nil {
-			return nil, err
-		}
-
-		return &appConfig, nil
-	}
-
-	var appConfig AppConfig
-
-	if err := json.Unmarshal(bytes, &appConfig); err != nil {
-		return nil, fmt.Errorf("Error unmarshalling config file: %s", err)
-	}
-
-	return &appConfig, nil
 }
